apps/api/pyproject.toml
[project]
name = "veritas-api"
version = "0.1.0"
description = "KYC/AML Automation Platform - Document extraction, sanctions screening, and risk scoring"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # API Framework
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.12",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Document Processing
    "pillow>=11.0.0",
    "pdf2image>=1.17.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    # OCR
    "pytesseract>=0.3.13",
    "opencv-python>=4.10.0",
    # MRZ Parsing
    "mrz>=0.6.2",
    "google-cloud-vision>=3.12.0",
    "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "aiosqlite>=0.20.0",
]
